pyproject.toml
[project]
name = "autodask"
version = "0.1.0"
description = "AutoML Based on Dask"
authors = [
    {name = "Dmitry Gilemkhanov",email = "dima.rize@yandex.ru"}
]
license = {text = "Apache-2.0 license"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dask[complete] (>=2025.4.1,<2026.0.0)",
    "dask-ml (>=2025.1.0,<2026.0.0)",
    "dask-xgboost (>=0.2.0,<0.3.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "catboost (>=1.2.8,<2.0.0)",
    "setuptools (>=80.3.1,<81.0.0)",
    "numpy (>=2.2.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"
